djangoapp/requirements.txt
Django>=4.2.1,<4.3
psycopg2-binary>=2.9.6,<2.10
Pillow